refactor: add `env.ts` separated from `index.ts`
